refactor: normalize components structure (plain sctructure for organisms)
